refactor: #4 🛠️ migrate to typescript
