+++ packages/dataset/index.ts
@@ -1,3 +1,3 @@
-export type * from "./types";
+export type * from "../types/index";
 
 export { KANJI_DATA_SET } from "./kanji";
