+++ packages/frontend/errors.ts
@@ -1,0 +1,17 @@
+export class InitializeError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
+
+export class CanvasCtxNotFoundError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
+
+export class CandidateContainerNotFoundError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
